requirements.txt
alembic==1.4.3
astroid==2.4.2
blinker==1.4
Bootstrap-Flask==1.5.1
click==7.1.2
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Avatars==0.2.2
Flask-CKEditor==0.4.4.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.0.1
PyJWT==1.7.1
pylint==2.6.0
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-editor==1.0.4
rope==0.18.0
six==1.15.0
SQLAlchemy==1.3.20
timeago==1.0.14
toml==0.10.2
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
